fix: R1 — wire comfort labels into ScoredHour (#276)
Extract pure comfort scoring from presenter to src/lib/outdoor-comfort.ts.
Add comfort: string field to ScoredHour, computed in score-hour.ts.
Presenter now delegates to lib for score + text, keeps styled label.
Pipeline dump includes comfort in editorial context.

11 new tests (lib comfort score + text). All 731 tests pass.
